small directory restructuring
